avoid duplicated code for vector_groupx
